Dotted-tail notation
Use this notation to write a procedure same-parity that takes one or more integers and returns a list of all the arguments that have the same even-odd parity as the first argument. 
Also defining a procedure select that takes a predicate and a list and returns a list consisting of all elements of the given list that fulfill the predicate.